tests: stabilize tests more
