+++ app/assets/json/translations/react/en.json
@@ -1576,6 +1576,7 @@
   "signIn.newAccount.p1": "We sent a link to %{email}.",
   "signIn.newAccount.p2": "Follow the link in your email to finish creating your account. It expires after 24 hours.",
   "signIn.newAccount.sendEmailAgainButton": "Send email again",
+  "signIn.newAccount.sendEmailAgainButton.confirmation": "Email sent. Check your email.",
   "t.ami": "AMI",
   "t.applicationClosed": "Applications Closed",
   "t.applicationDeadline": "Application Deadline",
